% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ChangeContinuousTraitInDiscrete}
\alias{ChangeContinuousTraitInDiscrete}
\title{Convert several continuous variables in discrete variables}
\usage{
ChangeContinuousTraitInDiscrete(Matrix, columnsIndex, Nstates, subclass)
}
\arguments{
\item{Matrix}{array matrix of continuous variables}

\item{columnsIndex}{vector of integers mentioning the columns to convert in integers}

\item{Nstates}{number of states of the defined traits (could be a vector or a value)}

\item{subclass}{intervals: states are fairly split. ordinal: ordered states,
split is random. non_eq_nominal: split is random, no order (shuffled)}
}
\value{
A data frame with the integer variables and continuous variables.
}
\description{
This function converts continuous variables in discrete variables. The variables can be converted as
interval, ordinal or nominal.
}
