% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConvertContinousInDiscreteValues}
\alias{ConvertContinousInDiscreteValues}
\title{Converts a single continuous variable in discrete variable}
\usage{
ConvertContinousInDiscreteValues(values, Nstates, subclass)
}
\arguments{
\item{values}{vector of float}

\item{Nstates}{number of states of the defined traits}

\item{subclass}{intervals: states are fairly split. ordinal: ordered states,
split is random. non_eq_nominal: split is random, no order (shuffled)}
}
\value{
A vector with integer instead of floats
}
\description{
This function converts a continuous variable in discrete variable. The variable can be converted as interval,
ordinal or nominal.
}
