% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertOneHotEncoIntoFactors}
\alias{convertOneHotEncoIntoFactors}
\title{Conversion one hot encoding in factors}
\usage{
convertOneHotEncoIntoFactors(oneHotData, nativeColNames)
}
\arguments{
\item{oneHotData}{data frame of one one hot encoding}

\item{nativeColNames}{columns names of the original column names of the one hot encoding(before the conversion in one hot
encoding).}
}
\value{
a data.frame in which each variable that were one hot encoding are now factors.
}
\description{
This function converts data frame of one hot encoding
representing a categorical variable.
}
