% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corContiTraitsOneTrait}
\alias{corContiTraitsOneTrait}
\title{Continuous traits correlated to a single trait discrete trait}
\usage{
corContiTraitsOneTrait(nbrTraits, discreteTrait, highCor, shuffle)
}
\arguments{
\item{nbrTraits}{vector, defining the number of traits (columns)}

\item{highCor}{numerical, strength of the correlation.}

\item{shuffle}{boolean, if true discrete states are shuffled, False, discrete states are not shuffled}

\item{discreteTraits}{vector from which the simulated traits are correlated}
}
\value{
A data frame of continuous traits where each column is a trait. Continuous traits are numeric.
}
\description{
This function generates a data frame of continuous traits
which are all correlated with a single traits but uncorrelated with each other.
}
