% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corMatrixConti}
\alias{corMatrixConti}
\title{Correlated continuous variable}
\usage{
corMatrixConti(discreteTrait, highCor, shuffle)
}
\arguments{
\item{discreteTrait}{vector of integers}

\item{highCor}{numerical, strength of the correlation.}

\item{shuffle}{boolean, if true discrete states are shuffled, False, discrete states are not shuffled}
}
\value{
A vector of continuous variables normally distributed and correlated to the discrete trait.
}
\description{
This function generates a continuous trait which is correlated to a discrete trait according to a
correlation factor. The function applies the function rnorm_pre from the R package faux.
}
