% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Empirical data and phylogeny in R list}
\usage{
data_preprocessing(empData, empTree = NULL)
}
\arguments{
\item{empData}{a table of class data.frame. Must have a column named Species}

\item{empTree}{a phylogenetic tree, by default empTree = NULL.}

\item{save}{path to save the data}
}
\value{
return a list which mimic the structure of the simulated data list. Contain at least the empirical data as data
.frame and the phylogenetic tree as phylo object. The continuous traits are log transformed.
}
\description{
This function saves an empirical dataset and a phylogeny in a list of the same structure like when the data
is simulated and converts the character and integer columns as factors.
}
