% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataframe}
\alias{dataframe}
\title{Input for the simulation of trait data}
\format{
A data frame of 11 columns and 4 rows:
\describe{
\item{nbr_traits}{Number of traits simulated with specific parameters.}
\item{class}{Type of traits, (continuous, non_eq_nominal(categorical) and ordinal).}
\item{model}{Evolutionary model (BM1, OU1, ARD, SYM, ER, Manual).}
\item{states}{Number of states for discrete traits, (if continuous set it to 1).}
\item{correlation}{Index corresponding to the group of simulated traits which are correlated or not to other traits.}
\item{uncorr_traits}{Among the "nbr_traits", it's the number of uncorrelated traits.}
\item{fraction_uncorr_traits}{Fraction among the "nbr_traits" which are uncorrelated.}
\item{lambda}{Pagel's lambda.}
\item{kappa}{Pagel's kappa.}
\item{highCor}{Correlation rate between the trait defined in "manTrait" and the simulated traits.}
\item{manTrait}{Index of the trait correlated with the trait(s) simulated according to the manual model.}
}
}
\source{
{Created in-house to serve as an example.}
}
\usage{
dataframe
}
\description{
Data frame which provides the structure of the data that the user wants to simulate.
}
\examples{
data(dataframe)
}
\keyword{datasets}
