% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateOneHotEncoVariables}
\alias{generateOneHotEncoVariables}
\title{Conversion factors in one hot encoding}
\usage{
generateOneHotEncoVariables(NaNData)
}
\arguments{
\item{NaNData}{data frame of one or several factors columns}
}
\value{
A list containing a data frame in which each discrete variable is converted as one hot encoding and vector of
characters.
}
\description{
This function converts discrete variables in one hot encoding using the function one_hot from the R package
mltools
}
