% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eigenvec.R
\name{get_eigenvec}
\alias{get_eigenvec}
\title{Eigenvectors calculations}
\usage{
get_eigenvec(tree, variance_fraction)
}
\arguments{
\item{tree}{phylogenetic tree of class "phylo"}

\item{variance_fraction}{variance_fraction minimum variance (\%) explained by the eigenvectors}
}
\value{
a data frame in which each column represents an eigenvector
}
\description{
This function calculates the eigenvectors of a phylo object
}
