% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTipsNA}
\alias{getTipsNA}
\title{Identify species for phylogenetic NAs}
\usage{
getTipsNA(Tree, MinTips)
}
\arguments{
\item{Tree}{Phylogeny}

\item{missingRates}{numerical vector corresponding to the rate of missing value to introduce in the data}
}
\value{
A vector containing the names of the species that should get NAs for their traits
}
\description{
This function returns the names of the species that should get NAs for their traits
}
