% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_voting.R
\name{hard_voting}
\alias{hard_voting}
\title{Hard voting}
\usage{
hard_voting(listOfdataframe)
}
\arguments{
\item{listOfdataframe}{list containing dataframes}
}
\value{
A dataframe
}
\description{
This function aggregates the output of several imputed dataframes selecting the most recurring value for
each row.
}
