% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_error.R
\name{imputation_error}
\alias{imputation_error}
\title{Calculate error of imputation}
\usage{
imputation_error(imputedData, trueData, missingData, imputationApproachesName, dataframe)
}
\arguments{
\item{imputedData}{array of imputed data}

\item{trueData}{array of true data}

\item{missingData}{array of data with missing values}

\item{imputationApproachesName}{character providing the name of the imputed appraoch used}

\item{dataframe}{dataframe given information about the data (type of variables, model of evolution, ....)}
}
\value{
return a data frame with in the first column the trait names and in the second the errors
}
\description{
This function calculates the RMSE for imputed continuous data, the absolute
error for imputed ordinal data and the Proportion of falsely classified entries (PFC) for the other
subcategories of discrete data
}
