% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_categorical_traits.R
\name{imputeOneDiscreteTrait}
\alias{imputeOneDiscreteTrait}
\title{Imputation of missing data for one discrete trait}
\usage{
imputeOneDiscreteTrait(trait, Data)
}
\arguments{
\item{missingData}{data frame of 1 factor column containing NAs}

\item{tree}{phylo object}
}
\value{
a data frame of 1 factor column with the NAs replaced by values.
}
\description{
This function imputes missing values of one vector according to a Markov chain model using the R function
corHMM from the corHMM package.
}
