% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN_phylo.R
\name{kNN_phylo}
\alias{kNN_phylo}
\title{non-parametric missing value imputation for mixed-type data by kNN(k-nearest neighbor algorithm)}
\usage{
kNN_phylo(missingData, k, numFun, catFun, variance_fraction = 0, tree, hint = NULL)
}
\arguments{
\item{missingData}{data frame of 1 or more columns containing NAs}

\item{k}{integer, number of nearest neighbors used}

\item{numFun}{numFun: function for aggregating the kNN in the case of a numerical variable}

\item{catFun}{catFun: function for aggregating the kNN in the case of a categorical variable}

\item{variance_fraction}{variance_fraction minimum variance (\%) explained by the eigenvectors}

\item{tree}{phylo object}

\item{hint}{data frame already imputed by a comprative methods (Rphylopars for continuous and corHMM for discrete)}
}
\value{
A data frame of 1 or more numeric columns with the NAs replaced by values + parameters used for the imputation
}
\description{
This function imputes missing data for continuous and discrete traits applying the kNN approach
}
