% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_miss_meca.R
\name{mar_miss_meca}
\alias{mar_miss_meca}
\title{Impute NaNs in a simulated data according the MAR mechanism}
\usage{
mar_miss_meca(missingRate, ds, cols_mis, cols_ctrl)
}
\arguments{
\item{missingRate}{numerical vector corresponding to the rate of missing value to introduce in the data}

\item{ds}{dataframe in which NA should be created}

\item{cols_mis}{vector of index or columns name where missing value will be created}

\item{cols_ctrl}{vector of index or columns name which control the creation of missing values (same size of cols_mis)}
}
\value{
a dataset with NA following a pattern of MAR.
}
\description{
Simulate NaNs in a complete data composed of discrete and continuous traits which are correlated
or uncorrelated. The NaNs are imputed according a missing rate and the missing mechanism MAR. It used the function delete_MAR
from the package missMethods.
}
