% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcar_miss_meca.R
\name{mcar_miss_meca}
\alias{mcar_miss_meca}
\title{Impute NaNs in a simulated data according the MCAR mechanism}
\usage{
mcar_miss_meca(missingRate, ds, cols_mis)
}
\arguments{
\item{missingRate}{numerical vector corresponding to the rate of missing value to introduce in the data}

\item{ds}{data frame in which NA should be created}

\item{cols_mis}{vector of index or columns name where missing values will be created}
}
\value{
a dataset with NA following a pattern of MCAR.
}
\description{
Simulate NaNs in a complete data composed of discrete and continuous traits which are
correlated or uncorrelated. The NaNs are imputed according a missing rate and the missing mechanism MCAR. It used the
function delete_MCAR from the package missMethods.
}
