% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_phylo.R
\name{mice_phylo}
\alias{mice_phylo}
\title{Imputation of missing values in mixed data by the MICE approach}
\usage{
mice_phylo(missingData, nbrMI, variance_fraction = 0, tree, hint = NULL)
}
\arguments{
\item{missingData}{data frame of 1 or more columns containing NAs}

\item{nbrMI}{integer, mentioning the total number of imputations}

\item{variance_fraction}{variance_fraction minimum variance (\%) explained by the eigenvectors}

\item{tree}{phylo object}

\item{hint}{data frame already imputed by a comparative methods (Rphylopars for continuous and corHMM for discrete)}
}
\value{
A data frame of 1 or more numeric columns with the NAs replaced by values + parameters used for the imputation
}
\description{
This function imputes missing data for continuous and discrete traits applying the MICE approach using the
predictive mean matching method for continuous traits, the polytomous regression for unordered categorical traits having
more than 2 states, if binary the algorithm uses a logistic regression.
}
