% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_imputation.R
\name{missing_data_imputation}
\alias{missing_data_imputation}
\title{Impute dataframe containing missing values}
\usage{
missing_data_imputation(ImputationApproachesNames, data, tree = NULL, strategies, varfrac = NULL, save = NULL)
}
\arguments{
\item{ImputationApproachesNames}{names of the imputations methods (character vector)}

\item{data}{dataframe containing NAs}

\item{tree}{phylo object}

\item{strategies}{character vector "NP", "P", "2-step"}

\item{maxit}{maxit maximum number of iteration (mice)}

\item{nbrMI}{integer, mentioning the total number of imputations (mice)}

\item{k}{integer, number of nearest neighbors used (kNN)}

\item{numFun}{numFun: function for aggregating the kNN in the case of a numerical variable}

\item{catFun}{catFun: function for aggregating the kNN in the case of a categorical variable}

\item{mincor}{minimal correlation for prediction matrix in mice}

\item{varfrac}{amount of variance explained by the eigenvectors of the phylogenetic tree (provide only if in strategy
there is "P")}

\item{save}{name of the output}
}
\value{
A list with the data imputed and all the parameters used for.
}
\description{
This function imputes the dataset containing missing values with Rphylopars, corHMM,
MICE, missForest, KNN and GAIN according to 3 strategies, without phylogeny, with phylogenetic informations and 2-step
}
