% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_insertion.R
\name{na_insertion}
\alias{na_insertion}
\title{Impute NaNs in a empirical data according to MCAR, MAR, MNAR and PhyloNa}
\usage{
na_insertion(missingRates, dataset, missingTraits, MARtrait, save = NULL)
}
\arguments{
\item{missingRate}{float corresponding to the rate of missing value to introduce in the data}

\item{dataset}{data.frame}

\item{missingTraits}{numerical, total number of traits in which missing data will be generated}

\item{MARTraits}{index or vector, variables in which the missing values will be generated. (only for MAR) same length
than MARctrlTraits}

\item{MARctrlTraits}{index or vector, variables on which the missing values will be designed. (only for MAR) same length
than MARTraits}

\item{traitsNoNA}{integer mentioning the columns in which we don't want NA generation.}

\item{tree}{phylo object, by default NULL (in case no phylogenetic tree)}

\item{save}{character correspond to the name of the saved file in .RData format}
}
\value{
A nested list composed of the partitioned data with the 4 categories of missing data (MCRA, MAR, MNAR and PhyloNa
) according to a precise missing rate and of list of the parameters used for the missing values generation
}
\description{
Simulate NaNs in a complete data set composed of continuous and discrete traits. The Nas are imputed according a
missing rate and respecting 4 categories of missing data, MCAR, MAR, MNAR and PhyloNa.
}
