% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloNa_miss_meca.R
\name{phyloNa_miss_meca}
\alias{phyloNa_miss_meca}
\title{Impute NaNs in a simulated data according the PhyloNa mechanism}
\usage{
phyloNa_miss_meca(missingRate, ds, tree)
}
\arguments{
\item{missingRate}{numerical vector corresponding to the rate of missing value to introduce in the data. Should be 0 < missing rate <= 0.5}

\item{ds}{dataframe in which NA should be created}

\item{tree}{object of class phylo}
}
\value{
A dataset with NA following a pattern of MNAR.
}
\description{
Simulate NaNs in a complete data composed of discrete and continuous traits which are correlated
or uncorrelated. The NaNs are imputed according a missing rate and the missing mechanism PhyloNa. It used the function
getTipsNA().
}
