% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_categorical_traits.R
\name{pi_categorical_traits}
\alias{pi_categorical_traits}
\title{Imputation of missing data in discrete traits, mandatory phylogenetic information}
\usage{
pi_categorical_traits(missingData, tree)
}
\arguments{
\item{missingData}{data.frame of 1 or more factor columns containing NAs}

\item{tree}{phylo object}
}
\value{
a data.frame of 1 or more factor columns with the NAs replaced by values, parameters used for the
imputation, matrix of probabilities of each states
}
\description{
This function applies the function imputeOneDiscretTrait on several columns at one time.
}
