% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescaleTree}
\alias{rescaleTree}
\title{Rescale a phylogenetic tree}
\usage{
rescaleTree(tree, subdata)
}
\arguments{
\item{tree}{phylo object}

\item{subdata}{a dataframe composed of the columns of the parameter called dataframe of the data_simulator function}
}
\value{
a phylo object rescaled according to the model define in the data frame.
}
\description{
This function rescales a phylogenetic tree of class phylo according to kappa or lambda (Pagel)
}
