% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simAlpha}
\alias{simAlpha}
\title{Simulate alpha matrix for morphological evolution}
\usage{
simSigma(Ntraits, alpha = NULL)
}
\arguments{
\item{Ntraits}{number of traits}

\item{alpha}{strength of attraction towards morphological optimum. Default between 0.5 and 2.
Optional, can be fixed to be equal between all traits by giving one value or
Ntraits*(Ntraits-1)/2}
}
\value{
matrix Ntrait x Ntrait for simulating trait evolution
}
\description{
This function generates an alpha matrix for
morphological evolution, defining the strength of attraction
towards a morphological optimum
}
