% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simDiscreteTraits}
\alias{simDiscreteTraits}
\title{Simulate discrete traits for all species according to the MK model}
\usage{
simDiscreteTraits(Ntraits, Nstates, rate_model, max_rate, SimTree)
}
\arguments{
\item{Ntraits}{number of traits}

\item{Nstates}{number of states of the defined traits (could be a vector or a value)}

\item{rate_model}{vector, rate model that the transition matrix must satisfy. Chose between "ER" (=all permitted
transitions occur at the same rate), "SYM" (=hat backward & forward transitions occur at the same rate) and "ARD" (=all
allowed transitions can occur at different rates).}

\item{tree}{stochastic birth-death trees}

\item{equal}{if TRUE, the transition matrix is equal for all the traits. If FALSE all the traits have a different
transition matrix. By default is TRUE.}

\item{Ordinal}{simulate ordinal data. Default is FALSE}
}
\value{
A list containing morphological evolution for each trait and each species, a matrix Ntrait x Ntrait for simulating
trait evolution and sigma matrix Ntraits x Ntraits
}
\description{
This function generates matrix of discrete values (states) of
morphological evolution. Define also a matrix of nodes
}
