% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulatedData}
\alias{simulatedData}
\title{Simulated trait data set}
\format{
A data frame of 13 columns and 100 rows:
\describe{
\item{I1.0/1}{Categorical trait simulated throught a MK model with a transition matrix obtained from an ARD rate model.}
\item{I2.0/1, I3.0/1, I4.0/1}{Categorical traits simulated manually in a way that one state is highly correlated to a state in "I1.0/1".}
\item{F1.0/3, F2.0/3, F3.0/3}{Continuous traits simulated throught a normal distribution and that are correlated at 80\% with "I1.0/1".}
\item{I1.0/4 I2.0/4 I3.0/4}{Categorical trait simulated throught a MK model with a transition matrix obtained from an ARD rate model. The 3 traits are independent.}
\item{F1.1/2, F2.1/2, F3.1/2}{Continuous traits simulated throught a brownian motion model(BM). The 3 traits are independent.}
}
}
\source{
{Created in-house to serve as an example.}
}
\usage{
simulatedData
}
\description{
Data frame composed of 13 traits and 100 species. All the discrete traits are composed of 3 states. The trait simulated thought a brownian motion model or an MK model, are generated according an ultrametric tree with a strong phylogenetic signal.
}
\examples{
data(simulatedData)
}
\keyword{datasets}
