% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_amputation.R
\name{state_amputation}
\alias{state_amputation}
\title{Amputation of values according the state and the continuous value}
\usage{
state_amputation(data, state, amount, columnNa, columnConti)
}
\arguments{
\item{data}{dataframe}

\item{state}{character: name of the state in which ampute data}

\item{amount}{numeric: amount of missing value within the state}

\item{columnNa}{integer: index of the column in which ampute data}

\item{columnConti}{integer: index of the column correlated to columnNa}

\item{level}{character: if top, ampute highest values. If low, remove lowest values. If rdn, remove randomly values}
}
\value{
A dataframe with missing value in the columnNa
}
\description{
Function to set NA for the same species in the random and top scenario
}
