% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_amputation_sameNA_top_rnd.R
\name{state_amputation_sameNA_top_rnd}
\alias{state_amputation_sameNA_top_rnd}
\title{Amputation of values according the state and the continuous value}
\usage{
state_amputation_sameNA_top_rnd(data, state, amount, columnNa, columnConti)
}
\arguments{
\item{data}{dataframe}

\item{state}{character: name of the state in which ampute data}

\item{amount}{numeric: amount of missing value within the state}

\item{columnNa}{integer: index of the column in which ampute data}

\item{columnConti}{integer: index of the column correlated to columnNa}
}
\value{
A dataframe with missing value in the columnNa
}
\description{
Function to set NA for the same species in the random and top scenario
}
