% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Simulated phylogenetic tree}
\format{
A phylo object composed of 100 tips with a birth rate of 0.4 and a death rate of 0.1.
}
\source{
{Created in-house to serve as an example.}
}
\usage{
tree
}
\description{
Ultrametric phylogenetic tree of class phylo simulated according a birth death process.
}
\examples{
data(tree)
}
\keyword{datasets}
