% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_amputation.R
\name{state_amputation}
\alias{state_amputation}
\title{Amputation of values according the state and the continuous value}
\usage{
state_amputation(data, state, amount, level, columnNa, columnConti)
}
\arguments{
\item{state}{character: name of the state in which ampute data}

\item{amount}{numeric: amount of missing value within the state}

\item{level}{character: if top, ampute highest values. If low, remove lowest values. If rdn, remove randomly values}

\item{columnNa}{integer: index of the column in which ampute data}

\item{columnConti}{integer: index of the column correlated to columnNa}

\item{data:}{dataframe}
}
\value{
A dataframe with missing value in the columnNa
}
\description{
This function amputes value according their states and the continuous
values of the correlated trait.
}
